adding cost model and address comments
